+++ inst/examples/ex-interval.R
@@ -1,0 +1,2 @@
+## HDR of the Old Faithful eruption times
+interval_hdr(faithful$eruptions)